pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coding-agent"
version = "0.1.0"
description = "Windows-first self-hosted coding agent with workspace guard, policy engine, memory, and CLI/API interfaces."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "pydantic-settings>=2.2.1,<3.0.0",
  "pyyaml>=6.0.1,<7.0.0",
  "typer>=0.12.3,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0,<9.0.0",
]

[project.scripts]
agent = "coding_agent.cli.main:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

